# Improve gc tricolor by avoiding revisiting nodes