@dmcgowan 2024-08-27
This is a good find and think your assessment is accurate. I think in the function description we already calls out that `func(ref Node) ([]Node, error)` should return the same results during the call, although we could make that more clear.